Style invite-users view and add helpers to handlebars
