+++ server/app.js
@@ -16,42 +16,36 @@ const api_router = require('./routers/api');
 const auth_router = require('./routers/auth');
 const admin_router = require('./routers/admin');
 
-const configDB = require('./config/database.js');
+const { https_redirect } = require('./custom_middleware');
+
+const configDB = require('./config/database');
+const passportInit = require('./config/passport');
 const helpers = require('./helpers.js');
 
 const app = express();
 
 mongoose.connect(configDB.url);
-require('./config/passport')(passport);
-
-const https_redirect = (req, res, next) => {
-  if (process.env.NODE_ENV === 'production') {
-    if (req.headers['x-forwarded-proto'] != 'https') {
-      return res.redirect(`https://${req.headers.host}${req.url}`);
-    } else {
-      return next();
-    }
-  } else {
-    return next();
-  }
-};
+passportInit(passport);
 
 app.use(https_redirect);
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 
-app.set('view engine', 'hbs');
+const hbs = exphbs.create({
+  extname: 'hbs',
+  layoutsDir: path.join(__dirname, 'views', 'layouts'),
+  partialsDir: path.join(__dirname, 'views', 'partials'),
+  defaultLayout: 'main',
+  helpers: {
+    setVariable: function(name, value) {
+      this[name] = value;
+    }
+  }
+});
+
 app.set('views', path.join(__dirname, 'views'));
-app.engine(
-  'hbs',
-  exphbs({
-    extname: 'hbs',
-    layoutsDir: path.join(__dirname, 'views', 'layouts'),
-    partialsDir: path.join(__dirname, 'views', 'partials'),
-    defaultLayout: 'main',
-    helpers: helpers
-  })
-);
+app.engine('hbs', hbs.engine);
+app.set('view engine', 'hbs');
 
 app.use(
   session({
@@ -63,6 +57,7 @@ app.use(
 app.use(passport.initialize());
 app.use(passport.session());
 app.use(flash());
+
 app.use(morgan('dev'));
 
 // Our Routers
