+++ server/views/layouts/main.hbs
@@ -9,7 +9,7 @@
   <link rel="stylesheet" href="style.min.css">
   <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,500,600,700" rel="stylesheet">
 </head>
-<body class="bg-near-white">
+<body class="bg-light-blue">
 
 {{{ body }}}
