+++ server/views/partials/name-email-invite-inputs.hbs
@@ -1,0 +1,9 @@
+{{setVariable "inputClass" "w-100 bg-light-blue br2 mb3 h2 ba b--black-20 tc f5 lh-copy fw4 pa3"}}
+<div class="flex flex-row">
+  <div class="w-third">
+    <input type="text" name="{{name}}" class="{{inputClass}}">
+  </div>
+  <div class="w-two-thirds">
+    <input type="email" name="{{email}}" class="{{inputClass}} ml3">
+  </div>
+</div>
